Add unit tests for Actor, Language, and Film services; validate CRUD operations and exception handling

+++ catalogo/src/test/java/cap/maria/catalogo/EntitiesTest/FilmCategoryPKTest.java
@@ -24,7 +24,7 @@ public class FilmCategoryPKTest {
 	void testParameterizedConstructor() {
 		FilmCategoryPK filmCategoryPK = new FilmCategoryPK();
 		filmCategoryPK.setFilmId(1);
-		filmCategoryPK.setCategoryId(2);
+		filmCategoryPK.setCategoryId((byte) 2);
 		assertAll("Constructor",
 			() -> assertEquals(1, filmCategoryPK.getFilmId(), "filmId"),
 			() -> assertEquals(2, filmCategoryPK.getCategoryId(), "categoryId"));
@@ -34,11 +34,11 @@ public class FilmCategoryPKTest {
 	void testEquals() {
 		FilmCategoryPK filmCategoryPK1 = new FilmCategoryPK();
 		filmCategoryPK1.setFilmId(1);
-		filmCategoryPK1.setCategoryId(2);
+		filmCategoryPK1.setCategoryId((byte) 2);
 
 		FilmCategoryPK filmCategoryPK2 = new FilmCategoryPK();
 		filmCategoryPK2.setFilmId(1);
-		filmCategoryPK2.setCategoryId(2);
+		filmCategoryPK2.setCategoryId((byte) 2);
 
 		assertTrue(filmCategoryPK1.equals(filmCategoryPK2));
 		assertTrue(filmCategoryPK2.equals(filmCategoryPK1));
@@ -48,11 +48,11 @@ public class FilmCategoryPKTest {
 	void testNotEquals() {
 		FilmCategoryPK filmCategoryPK1 = new FilmCategoryPK();
 		filmCategoryPK1.setFilmId(1);
-		filmCategoryPK1.setCategoryId(2);
+		filmCategoryPK1.setCategoryId((byte) 2);
 
 		FilmCategoryPK filmCategoryPK2 = new FilmCategoryPK();
 		filmCategoryPK2.setFilmId(2);
-		filmCategoryPK2.setCategoryId(3);
+		filmCategoryPK2.setCategoryId((byte) 3);
 
 		assertFalse(filmCategoryPK1.equals(filmCategoryPK2));
 		assertFalse(filmCategoryPK2.equals(filmCategoryPK1));
@@ -62,11 +62,11 @@ public class FilmCategoryPKTest {
 	void testHashCode() {
 		FilmCategoryPK filmCategoryPK1 = new FilmCategoryPK();
 		filmCategoryPK1.setFilmId(1);
-		filmCategoryPK1.setCategoryId(2);
+		filmCategoryPK1.setCategoryId((byte) 2);
 
 		FilmCategoryPK filmCategoryPK2 = new FilmCategoryPK();
 		filmCategoryPK2.setFilmId(1);
-		filmCategoryPK2.setCategoryId(2);
+		filmCategoryPK2.setCategoryId((byte) 2);
 
 		assertEquals(filmCategoryPK1.hashCode(), filmCategoryPK2.hashCode());
 	}
@@ -75,11 +75,11 @@ public class FilmCategoryPKTest {
 	void testHashCodeNotEquals() {
 		FilmCategoryPK filmCategoryPK1 = new FilmCategoryPK();
 		filmCategoryPK1.setFilmId(1);
-		filmCategoryPK1.setCategoryId(2);
+		filmCategoryPK1.setCategoryId((byte) 2);
 
 		FilmCategoryPK filmCategoryPK2 = new FilmCategoryPK();
 		filmCategoryPK2.setFilmId(2);
-		filmCategoryPK2.setCategoryId(3);
+		filmCategoryPK2.setCategoryId((byte) 3);
 
 		assertNotEquals(filmCategoryPK1.hashCode(), filmCategoryPK2.hashCode());
 	}
@@ -88,7 +88,7 @@ public class FilmCategoryPKTest {
 	void testGettersAndSetters() {
 		FilmCategoryPK filmCategoryPK = new FilmCategoryPK();
 		filmCategoryPK.setFilmId(1);
-		filmCategoryPK.setCategoryId(2);
+		filmCategoryPK.setCategoryId((byte) 2);
 
 		assertEquals(1, filmCategoryPK.getFilmId());
 		assertEquals(2, filmCategoryPK.getCategoryId());
